requirements.txt
# ── Core ──────────────────────────────────────────────────────────────────────
pymongo>=4.7
sentence-transformers>=2.7
python-dotenv>=1.0

# ── Auth ──────────────────────────────────────────────────────────────────────
fastapi-users[beanie]>=13.0
motor>=3.3

# ── Web API ───────────────────────────────────────────────────────────────────
fastapi>=0.111
uvicorn[standard]>=0.29
python-multipart>=0.0.9

# ── File format support ────────────────────────────────────────────────────────
pypdf>=4.0
python-docx>=1.1
beautifulsoup4>=4.12

# ── Re-ranker ─────────────────────────────────────────────────────────────────
# sentence-transformers already listed above (CrossEncoder included)

# ── Payments ──────────────────────────────────────────────────────────────────
requests>=2.31

# ── Production ────────────────────────────────────────────────────────────────
gunicorn>=21.0